third

+++ src/main/java/com/sesame/ecommerceproject/Repository/IlingneCommandeRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IlingneCommandeRepository extends JpaRepository<LigneCommande, Integer> {
+public interface IlingneCommandeRepository extends JpaRepository<LigneCommande, Long> {
 }
